Move 'useObservedWidth' helper from 'breadcrumbs/util' to general '/util'
